+++ src/components/index.ts
@@ -1,4 +1,4 @@
 export { default as Input } from '@/components/@common/Input';
 export { default as Search } from '@/components/Search';
-export { default as PeopleChoose } from '@/components/Search/PeopleChoose';
-export { default as PeopleCounter } from '@/components/Search/PeopleCounter';
+export { default as GuestCountInput } from '@/components/Search/GuestCountInput';
+export { default as GuestCounter } from '@/components/Search/GuestCounter';
